requirement.txt
# Data Handling
pandas>=2.0.0
numpy>=1.24.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Deep Learning Framework (PyTorch)
# Note: This installs the CPU version by default in some environments. 
# For GPU support, usually you run a specific command, but this is the standard package.
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu126

# Progress Bars & Visualization
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn

# Machine Learning Utilities (Metrics)
scikit-learn>=1.3.0
opencv-python


#Backend
fastapi 
uvicorn 
python-multipart